Add kagglesdk (#623)
This is generated code. To the extent that it has been tested, it is
used in some of the public API code, which is exercised by the unit
tests.

It might be nice if there was a way to allow both this repo and
`kagglehub` to share this package without publishing to pypi.