Clean up errors (#639)
There were some errors that had been present a year or so. They are
gone, as are all the unused import statements.

I'm getting really frustrated by the formatter. Everything here was
formatted by `yapf` 0.40.2 but I must have merged something that wasn't
properly formatted because there is a large number of format changes.
Unfortunately, that was also true before I re-ran the formatter. I
wonder if maybe `yapf` isn't deterministic. (Edit: It seems to be
deterministic.)

There were two methods that didn't have a `self` parameter. There was a
`unicode` statement in a try-block to test for Python 2 vs 3. There were
two undefined vars -- one was misspelled ('query' should have been
'quiet') and the other was a missing `path` parameter to the method. I
think that is all the errors I fixed.

Added #640 to track the lack of unit test coverage.