Add kaggle_api.py (#674)
This file is used in a bit less than 4000 notebooks. I've tried to remap
the functions to their CLI equivalents.

Some of the functions originally in this file separated file upload from
some other function, like dataset creation. Since the CLI functions
manage the upload, I did not implement those functions here, but left
instructions on what to use instead. Not ideal, but we don't want the
files getting uploaded twice.

This is completely untested, and I hope to fix that soon. The main items
that need testing are those that write a metadata file, or have a large
number of parameters.