Update a file that was omitted previously (#686)
When I last generated `kagglesdk` I apparently forgot to replace the one
in `src`. This file had a lot of new content and it needs to be added to
the one in `src`. This code is generated by `kapigen`.

Yeah, source management is awkward. I plan to fix that, too.