Add type annotations for mypy (#746)
I moved `FileList` from the bottom of the file to the top.

I introduced a number of typed variables to represent the return type.
Not sure why, but `mypy` claimed they had type `Any` otherwise. This is
mostly in the response objects, which are defined in `kagglesdk`, and I
don't type-check that library because it has a lot of errors. That may
have something to do with it.

The `lookup_enum()` method exists for legacy code/backward
compatibility. To eliminate a bunch of casts, I added a parameter that
is a sample value of the enum that needs to have a value extracted.

Type checking uncovered some errors in code I haven't tested. I kinda
had to fix them to move on. Likewise, I noticed a number of docstring
comments that needed some work, so I did it.