Update kagglesdk with dataset_delete endpoint, etc (#761)
@jplotts This has my latest with some of yours, too. No need to review,
unless you want to.

The change to the http client is from Auto-Kaggle. I updated kapigen
earlier, but not here.

And yes, some day (soonish) I need to start using the official
`kagglesdk` repo. That's going to require more mucking about in the
build and setup scripts, so I keep putting it off.