Rename "yes" params and make confirmation consistent (#765)
This was more involved than I expected. I noticed that `confirmation()`
wasn't as informative as the generated code confirmation, but I thought
a unified mechanism would be better. And it didn't make sense to
`exit()` if the response was "no". If the confirmation had been done in
the `_cli` then that would be OK, but it isn't. At least not today.

I haven't tested the `_cli` methods yet. I'm working on a PR (#766) to
do so, but it needs to run against prod, so I need to deploy this first.