Clean up project config files and version (#788)
This moves the `hatch` version file to be part of `kaggle` so we can use
the `__version__`. Now, instead of having to update the version in many
files, we just do `hatch version ...` then update the change log. I'll
update the maintainer docs, too. This could break some code, but it
seems low risk.

It also deletes obsolete `setuptools` files.