Reformat kagglesdk to make strings quotes consistent (#802)
I accidentally introduced double-quotes to the codebase with my previous
formatting PR. Due to `skip-string-normalization`, `black` did not make
any changes to string quotes. I don't have a strong opinion on single vs
double but it is better to be consistent and let the formatter enforce
it.

Command to format the whole repo (after `pip3 install black`):
```bash
$ find . -name "*.py" | xargs black
```