Add sphjinx config files and update some docs (#814)
The `sphinx` config is a work-in-progress, but it does generate
good-looking docs now.
```bash
cd documentation
make html
```
Then open `documentation/_build/index.html' in the browser.

I had Gemini CLI condense the `GEMINI.md` file and decided to add it
here since it is now fairly generic (that is, not specific to me).

I noticed one of the tutorials was formatted differently than the others
so I changed it for consistency.

Most of the new files in `documentation` are for `sphinx`. `intro.md`
isn't new but I had a git mistake and it looks new. I just added
sections on installing `kaggle` and setting up the API key. The stuff in
`documentation/source` comes from `sphinx` and we aren't using it.