Give Gemini CLI some instruction about analyzing Python (#824)
Sometimes, when asking Gemini CLI to find docstrings it constructs a
Python function that is too simple to find them all. This instructions
it to use `api.get_docstrting()`. When working with type hints it seems
to also be a bit simplistic so I'm trying to get it to use tools in
`ast'. Not sure if the latter actually makes sense, though, in the
context of vibe-coding.