Add type hints to docstrings (#827)
When I was reviewing the generated docstrings I found myself wondering
what types all the args had. This adds types to args in the generated
docs (it should, have yet to verify). But, having done this, I'm
wondering if it is such a good idea. We now have duplicate type decls
for each method, which need to be kept in sync. OTOH, the docs can be
read w/o looking at the code to find the types.